refactor(webui): extract archive-after duration helpers to lib
The Create Task screen and the routing-rule editor both convert between
a whole-hours Select value and a protobuf Duration. Move
durationFromHours / hoursFromDuration into a shared lib/duration module
so the two screens share one implementation instead of duplicating it.
